+++ src/main/kotlin/me/rafaelldi/aspire/run/AspireHostSettingsEditor.kt
@@ -18,6 +18,7 @@ class AspireHostSettingsEditor(private val project: Project) :
             lifetime,
             project.runnableProjectsModelIfAvailable,
             ProjectSelector(AspireBundle.message("run.editor.project"), "Project"),
+            LaunchProfileSelector(AspireBundle.message("run.editor.launch.profile"), "Launch_profile"),
             EnvironmentVariablesEditor(AspireBundle.message("run.editor.environment.variables"), "Environment_variables"),
             ViewSeparator(AspireBundle.message("run.editor.open.browser")),
             TextEditor(AspireBundle.message("run.editor.url"), "URL", lifetime),
@@ -27,20 +28,18 @@ class AspireHostSettingsEditor(private val project: Project) :
     }
 
     override fun applyEditorTo(configuration: AspireHostConfiguration) {
-        val selectedProject = viewModel.projectSelector.project.valueOrNull
-        if (selectedProject != null) {
-            configuration.parameters.apply {
-                projectFilePath = selectedProject.projectFilePath
-                trackEnvs = viewModel.trackEnvs
-                envs = viewModel.environmentVariablesEditor.envs.value
-                trackUrl = viewModel.trackUrl
-                startBrowserParameters.url = viewModel.urlEditor.text.value
-                startBrowserParameters.browser = viewModel.dotNetBrowserSettingsEditor.settings.value.myBrowser
-                startBrowserParameters.startAfterLaunch =
-                    viewModel.dotNetBrowserSettingsEditor.settings.value.startAfterLaunch
-                startBrowserParameters.withJavaScriptDebugger =
-                    viewModel.dotNetBrowserSettingsEditor.settings.value.withJavaScriptDebugger
-            }
+        configuration.parameters.apply {
+            projectFilePath = viewModel.projectSelector.project.valueOrNull?.projectFilePath ?: ""
+            profileName = viewModel.launchProfileSelector.profile.valueOrNull?.name ?: ""
+            trackEnvs = viewModel.trackEnvs
+            envs = viewModel.environmentVariablesEditor.envs.value
+            trackUrl = viewModel.trackUrl
+            startBrowserParameters.url = viewModel.urlEditor.text.value
+            startBrowserParameters.browser = viewModel.dotNetBrowserSettingsEditor.settings.value.myBrowser
+            startBrowserParameters.startAfterLaunch =
+                viewModel.dotNetBrowserSettingsEditor.settings.value.startAfterLaunch
+            startBrowserParameters.withJavaScriptDebugger =
+                viewModel.dotNetBrowserSettingsEditor.settings.value.withJavaScriptDebugger
         }
     }
 
@@ -48,6 +47,7 @@ class AspireHostSettingsEditor(private val project: Project) :
         configuration.parameters.apply {
             viewModel.reset(
                 projectFilePath,
+                profileName,
                 trackEnvs,
                 envs,
                 trackUrl,